start working on log passwords and log success status parsing
These changes are still not visible in the GUI as it is
work-in-progress. In fact I am currently unable to test the
implementation without impacting the real Opencaching.de site (spamming
the caches there while testing does not quite feel right), so further
changes on this will probably be delayed.