re-enable updates; show about dialog; work-around for Java 8
Some APIs are deprecated starting with Java 10, but we are currently
targeting at least Java 8. Unfortunately the new APIs are not available
in Java 8 for now. To keep the work on the user side as small as
possible, a compatibility version has been introduced. Nevertheless I
recommend to update to the latest Java version anyway, as this
compatibility layer will probably be removed as soon as the deprecated
methods have been removed from the latest Java version as well.